requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Django==5.0.6
django-types==0.19.1
pycodestyle==2.11.1
PyJWT==2.8.0
redis==5.0.4
sqlparse==0.5.0
types-psycopg2==2.9.21.20240417
